apps/orchestrator/requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0

# Temporal
temporalio>=1.5.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP client
httpx>=0.26.0
aiohttp>=3.9.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-json-logger>=2.0.7
